Stairs form submission, altar scanner
The stair scan will now send a message to backround.js with the level
and the stair location, and background.js will make a get request to our
form, passing the response status back to runner.

There's an altar scanner now too. Altar locations will be printed to the
chat if their location is different than that of the last altar.