Interval worker change, RAF spoofing, canvas hiding
Interval workers will now deactivate when given an interval of 0.

A new interval worker, rafWorker, will be created at start up. Its
onmessage calls jv.frame if auto-run, auto-attack, or the explo macro
are running. It is not activated at this point. Event listeners for blur
and focus have been added that will start and stop this worker,
respectively. Basically, if the tab loses focus I simulate animation
frames. This means that the explo macro can be used when the tab is
inactive.

Also on blur and focus the canvas element will be hidden/shown.
Animation buildup already isn't a problem with simulated animation
frames, but this allows multiple instances of the game to be run without
taxing the gpu.